reorganize fsal to be more self-contained
